requirements.txt
opencv-python
numpy
pyserial
pyrealsense2
